chore: rename controller -> gateway
