+++ cmd/server.go
@@ -41,7 +41,7 @@ import (
 	_ "github.com/flanksource/incident-commander/catalog"
 	_ "github.com/flanksource/incident-commander/connection"
 	_ "github.com/flanksource/incident-commander/playbook"
-	_ "github.com/flanksource/incident-commander/plugin/controller"
+	_ "github.com/flanksource/incident-commander/plugin/gateway"
 	_ "github.com/flanksource/incident-commander/shorturl"
 	_ "github.com/flanksource/incident-commander/snapshot"
 	"github.com/flanksource/incident-commander/teams"
