+++ plugin/gateway/invocation_audit.go
@@ -1,4 +1,4 @@
-package controller
+package gateway
 
 import (
 	"crypto/sha256"
