+++ plugin/gateway/proxy_test.go
@@ -1,4 +1,4 @@
-package controller
+package gateway
 
 import (
 	"net/http"
