+++ plugin/gateway/suite_test.go
@@ -1,4 +1,4 @@
-package controller
+package gateway
 
 import (
 	"testing"
